state: Reorganize files by stored information
Instead of read/write separation,
separate by what's stored in each file:
remote, repository information, branches, etc.
