+++ internal/spice/state/repo_test.go
